[DDD light] Refactored the board system to work with thw new architecture

+++ backend/ArhiTodo.Application/DTOs/Board/BoardGetDto.cs
@@ -4,7 +4,7 @@ namespace ArhiTodo.Application.DTOs.Board;
 
 public class BoardGetDto
 {
-    public int BoardId { get; set; }
+    public long BoardId { get; set; }
         
     public required string BoardName { get; set; }
